pyproject.toml
[project]
name = "kaggle-openai-to-z-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph (>=0.4.8,<0.5.0)",
    "langchain-openai (>=0.3.24,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "boto3 (>=1.38.41,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "rasterio (>=1.4.3,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pyproj (>=3.7.1,<4.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "kaggle_openai_to_z_challenge", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
